+++ lib/cenatus_ltd_web/templates/layout/_head.html.heex
@@ -17,6 +17,6 @@
   <link rel="stylesheet" href="https://fonts.googleapis.com/css?family=Open+Sans:300" />
   <link rel="stylesheet" href="https://cdn.jsdelivr.net/npm/bootstrap@5.2.3/dist/css/bootstrap.min.css"  integrity="sha384-rbsA2VBKQhggwzxH7pPCaAqO46MgnOM80zW1RWuH61DGLwZJEdK2Kadq2F9CUG65" crossorigin="anonymous">
 
-  <link rel="stylesheet" href={ Routes.static_path(@conn, "/css/app.css") } />
+  <link phx-track-static rel="stylesheet" href={ Routes.static_path(@conn, "/assets/from_scss.css")} />
 
   <%= csrf_meta_tag() %>
