+++ lib/cenatus_ltd_web/templates/layout/_scripts.html.heex
@@ -1,11 +1,11 @@
-<script src='https://code.jquery.com/jquery-2.2.4.js'></script>
+<script src='https://code.jquery.com/jquery-3.7.0.js'></script>
 <script src='https://cdnjs.cloudflare.com/ajax/libs/gsap/1.19.1/TweenMax.min.js'></script>
 
 <script type='text/javascript'>
   <%= raw("const mix_env = '#{Mix.env}';") %>
 </script>
 
-<script src={Routes.static_path(@conn, "/js/app.js")}></script>
+<script src={Routes.static_path(@conn, "/assets/app.js")}></script>
 
 <script type='text/javascript'>
   var gaJsHost = (("https:" == document.location.protocol) ? "https://ssl." : "http://www.");
